backend/requirements.txt
asgiref==3.8.1
astroid==3.3.6
attrs==23.2.0
Automat==22.10.0
autopep8==2.3.1
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
debugpy==1.8.11
dill==0.3.9
distro==1.9.0
distro-info==1.7+build1
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eralchemy==1.5.0
graphviz==0.20.3
greenlet==3.1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==1.1.1
isort==5.13.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
netifaces==0.11.0
oauthlib==3.2.2
packaging==24.0
pep8==1.7.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.12.1
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pylint==3.3.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
python-apt==2.7.7+ubuntu4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
systemd-python==235
tomlkit==0.13.2
Twisted==24.3.0
typing_extensions==4.10.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
Werkzeug==3.1.3
wheel==0.42.0
xkit==0.0.0
zope.interface==6.1
